backend/apps/portfolios/serializers.py
# apps/portfolios/serializers.py
from rest_framework import serializers
from .models import Portfolio

class PortfolioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Portfolio
        fields = ['id', 'title', 'description', 'image', 'video', 'link', 'created_at']