backend/requirements.txt
django 
psycopg2 
djangorestframework
django.core.management
djangorestframework-authtoken
Pillow